+++ deviceinfo.js
@@ -234,6 +234,9 @@ export default {
   getIPAddress: function() {
     return RNDeviceInfo.getIpAddress();
   },
+  getCameraPresence: function() {
+    return RNDeviceInfo.getCameraPresence();
+  },
   getMACAddress: function() {
     return RNDeviceInfo.getMacAddress();
   },
@@ -381,5 +384,11 @@ export default {
   },
   getSystemAvailableFeatures: function() {
     return RNDeviceInfo.getSystemAvailableFeatures();
+  },
+  isLocationEnabled: function() {
+    return RNDeviceInfo.isLocationEnabled();
+  },
+  getAvailableLocationProviders: function() {
+    return RNDeviceInfo.getAvailableLocationProviders();
   },
 };