+++ refresh-example.sh
@@ -3,6 +3,7 @@ set -e
 
 echo "You should run this from directory where you have cloned the react-native-device-info repo"
 echo "You should only do this when your git working set is completely clean (e.g., git reset --hard)"
+echo "You must run \`npm install react-native-cli -g\` prior to running this script"
 echo "You must have already run \`npm install\` in the repository so \`npx react-native\` will work"
 echo "This scaffolding refresh has been tested on macOS, if you use it on linux, it might not work"
 
@@ -15,7 +16,7 @@ else
   mkdir -p TEMP/android
   mkdir -p TEMP/ios
   cp example/README.md TEMP/
-  cp example/android/local.properties TEMP/android/
+  cp example/android/local.properties TEMP/android/ || true
   cp example/App.js TEMP/
   cp example/ios/Podfile TEMP/ios/
 fi
@@ -34,7 +35,7 @@ sed -i -e 's/ext {$/ext {        googlePlayServicesVersion = "16.1.0"/' android/
 rm -f android/build.gradle??
 
 # Patch the AndroidManifest directly to add our permissions
-sed -i -e 's/INTERNET" \/>/INTERNET" \/><uses-permission android:name="android.permission.ACCESS_WIFI_STATE" \/><uses-permission android:name="android.permission.BLUETOOTH" \/><uses-permission android:name="android.permission.READ_PHONE_STATE" \/>/' android/app/src/main/AndroidManifest.xml
+sed -i -e 's/INTERNET" \/>/INTERNET" \/><uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" \/><uses-permission android:name="android.permission.ACCESS_WIFI_STATE" \/><uses-permission android:name="android.permission.BLUETOOTH" \/><uses-permission android:name="android.permission.READ_PHONE_STATE" \/>/' android/app/src/main/AndroidManifest.xml
 rm -f android/app/src/main/AndroidManifest.xml??
 
 # Patch the AppDelegate for iOS battery level